requirements.txt
google-api-python-client
google-cloud
google-auth
google-cloud-bigtable
google-cloud-logging
google-cloud-monitoring
pyyaml

